@@ -26,13 +26,13 @@
          }
     },
     "require": {
-        "php": ">=7.3",
+        "php": ">=7.4",
         "comodojo/foundation": "dev-master",
         "ext-zip": "*"
     },
     "require-dev": {
         "phpunit/phpunit": "^9.0",
-        "scrutinizer/ocular": "^1.0,<1.7"
+        "scrutinizer/ocular": "^1.8"
     },
     "minimum-stability": "dev"
 }
